get rid of adaptor's operators; inline sources
